Move the code around a bit
